Treat user numbers a 32-bit signed integers, everywhere
we had mix of signed/unsigned and 16 and 32-bit integer representations of the
user number.

I somehow ended up with a 4GB (4294967296 byte) user.tab file today and don't
have an exact smoking gun but the conversions (mainly in userdat.c) between
unsigned 16-bit and signed 32-bit representation of the user number had me
suspicious.

Perform some userbase data file and index file sanity checking upon startup
of the terminal server thread and fail with critical error if anything seems
amiss (indication of corruption).

+++ src/sbbs3/chk_ar.cpp
@@ -306,7 +306,7 @@ bool sbbs_t::ar_exp(const uchar **ptrptr, user_t* user, client_t* client)
 				}
 				break;
 			case AR_USER:
-				if ((equal && user->number != i) || (!equal && user->number < i))
+				if ((equal && user->number != (int)i) || (!equal && user->number < (int)i))
 					result = _not;
 				else
 					result = !_not;
